Move files to internal namespace
