+++ firebaseai/src/Imagen/ImagenModel.cs
@@ -94,7 +94,7 @@ namespace Firebase.AI
           Firebase.AI.Internal.HttpHelpers.GetURL(_firebaseApp, _backend, _modelName) + ":predict");
 
       // Set the request headers
-      await Firebase.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
+      await Firebase.Internal.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
 
       // Set the content
       string bodyJson = MakeGenerateImagenRequest(prompt);
@@ -105,7 +105,7 @@ namespace Firebase.AI
 #endif
 
       var response = await _httpClient.SendAsync(request, cancellationToken);
-      await Firebase.HttpHelpers.ValidateHttpResponse(response);
+      await Firebase.Internal.HttpHelpers.ValidateHttpResponse(response);
 
       string result = await response.Content.ReadAsStringAsync();
 
@@ -203,7 +203,7 @@ namespace Firebase.AI
           Firebase.AI.Internal.HttpHelpers.GetTemplateURL(_firebaseApp, _backend, templateId) + ":templatePredict");
 
       // Set the request headers
-      await Firebase.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
+      await Firebase.Internal.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
 
       // Set the content
       Dictionary<string, object> jsonDict = new()
@@ -218,7 +218,7 @@ namespace Firebase.AI
 #endif
 
       var response = await _httpClient.SendAsync(request, cancellationToken);
-      await Firebase.HttpHelpers.ValidateHttpResponse(response);
+      await Firebase.Internal.HttpHelpers.ValidateHttpResponse(response);
 
       string result = await response.Content.ReadAsStringAsync();
