+++ firebaseai/src/TemplateGenerativeModel.cs
@@ -111,7 +111,7 @@ namespace Firebase.AI
           Firebase.AI.Internal.HttpHelpers.GetTemplateURL(_firebaseApp, _backend, templateId) + ":templateGenerateContent");
 
       // Set the request headers
-      await Firebase.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
+      await Firebase.Internal.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
 
       // Set the content
       string bodyJson = MakeGenerateContentRequest(inputs, chatHistory);
@@ -122,7 +122,7 @@ namespace Firebase.AI
 #endif
 
       var response = await _httpClient.SendAsync(request, cancellationToken);
-      await Firebase.HttpHelpers.ValidateHttpResponse(response);
+      await Firebase.Internal.HttpHelpers.ValidateHttpResponse(response);
 
       string result = await response.Content.ReadAsStringAsync();
 
@@ -142,7 +142,7 @@ namespace Firebase.AI
           Firebase.AI.Internal.HttpHelpers.GetTemplateURL(_firebaseApp, _backend, templateId) + ":templateStreamGenerateContent?alt=sse");
 
       // Set the request headers
-      await Firebase.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
+      await Firebase.Internal.HttpHelpers.SetRequestHeaders(request, _firebaseApp);
 
       // Set the content
       string bodyJson = MakeGenerateContentRequest(inputs, chatHistory);
@@ -153,7 +153,7 @@ namespace Firebase.AI
 #endif
 
       var response = await _httpClient.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
-      await Firebase.HttpHelpers.ValidateHttpResponse(response);
+      await Firebase.Internal.HttpHelpers.ValidateHttpResponse(response);
 
       // We are expecting a Stream as the response, so handle that.
       using var stream = await response.Content.ReadAsStreamAsync();
